refactor: Change Integer Id To Long For CustomPolygon

+++ sfo/src/main/java/com/graphhopper/sfo/custompolygon/dto/CustomPolygon.java
@@ -12,7 +12,7 @@ public class CustomPolygon {
 
     public CustomPolygon(){}
 
-    public CustomPolygon(int id, String geometry) {
+    public CustomPolygon(long id, String geometry) {
         this.id = id;
         this.geometry = geometry;
     }
